consolidate reusable functions in github-actions to utils folder
